+++ src/resizer/ReSizer.tsx
@@ -1,8 +1,8 @@
-import { Fragment } from 'react';
-import FlexResizer, {
-    getFlexSizeSetting, ResizerKindType,
-} from './FlexResizer';
+import { Fragment, useState } from 'react';
+import { getSetting, setSetting } from '../helper/settingHelper';
+import FlexResizer, { ResizerKindType } from './FlexResizer';
 
+export type Size = { [key: string]: string };
 export default function ReSizer({
     settingName,
     flexSizeDefault,
@@ -11,18 +11,22 @@ export default function ReSizer({
     children,
 }: {
     settingName: string,
-    flexSizeDefault: { [key: string]: string },
+    flexSizeDefault: Size,
     resizerKinds: ResizerKindType[],
     sizeKeys: [string, string, any?][],
     children: any[],
 }) {
-    const flexSize = getFlexSizeSetting(settingName, flexSizeDefault);
+    const [flexSize, setFlexSize] = useState(getFlexSizeSetting(settingName, flexSizeDefault));
     return (
         <>
             {sizeKeys.map(([key, classList, style = {}], i) => {
                 return (
                     <Fragment key={i}>
-                        {i !== 0 && <FlexResizer settingName={settingName}
+                        {i !== 0 && <FlexResizer
+                            checkSize={() => {
+                                const size = saveSize(settingName);
+                                setFlexSize(size);
+                            }}
                             type={resizerKinds[i - 1]} />}
                         <div data-fs={key} data-fs-default={flexSize[key]}
                             className={classList}
@@ -34,4 +38,27 @@ export default function ReSizer({
             })}
         </>
     );
+}
+
+const saveSize = (settingName: string) => {
+    const size: Size = {};
+    const rowItems: HTMLDivElement[] = Array.from(document.querySelectorAll('[data-fs]'));
+    rowItems.forEach((item) => {
+        size[item.getAttribute('data-fs') as string] = item.style.flex;
+    });
+    setSetting(settingName, JSON.stringify(size));
+    return size;
+};
+
+function getFlexSizeSetting(settingName: string, defaultSize: Size): Size {
+    const sizeStr = getSetting(settingName);
+    try {
+        const size = JSON.parse(sizeStr);
+        if (Object.keys(defaultSize).every((k) => size[k] !== undefined)) {
+            return size;
+        }
+    } catch (error) {
+        setSetting(settingName, JSON.stringify(defaultSize));
+    }
+    return defaultSize;
 }
